10.1 1:merge easyui版本修改 进入mysql

+++ views/testcase_report.py
@@ -278,7 +278,8 @@ def report_delete(testcase_time_id):
                 db.session.commit()
             from logs.config import path
             filename = os.path.join(path, testcase_report.filename)
-            os.remove(filename)
+            if testcase_report.filename:
+                os.remove(filename)
         except FileNotFoundError:
             pass
         FrontLogs('删除测试报告 id: %s 成功' % testcase_time_id).add_to_front_log()
@@ -346,5 +347,3 @@ testcase_report_blueprint.add_url_rule('/testcasereportdownload/<path:name>', vi
 
 testcase_report_blueprint.add_url_rule('/testcase_report_sendmail/', view_func=TestCaseReportSendMail.as_view('testcase_report_sendmail'))
 testcase_report_blueprint.add_url_rule('/email_validate/', view_func=ConfigMailValidate.as_view('/email_validate/'))
-
-
